+++ pkg/apis/options/options.go
@@ -52,17 +52,16 @@ type Options struct {
 
 	Providers Providers `cfg:",internal"`
 
-	APIRoutes             []string `flag:"api-route" cfg:"api_routes"`
-	SkipAuthRegex         []string `flag:"skip-auth-regex" cfg:"skip_auth_regex"`
-	SkipAuthRoutes        []string `flag:"skip-auth-route" cfg:"skip_auth_routes"`
-	SkipJwtBearerTokens   bool     `flag:"skip-jwt-bearer-tokens" cfg:"skip_jwt_bearer_tokens"`
-	ExtraJwtIssuers       []string `flag:"extra-jwt-issuers" cfg:"extra_jwt_issuers"`
-	SkipProviderButton    bool     `flag:"skip-provider-button" cfg:"skip_provider_button"`
-	SSLInsecureSkipVerify bool     `flag:"ssl-insecure-skip-verify" cfg:"ssl_insecure_skip_verify"`
-	SkipAuthPreflight     bool     `flag:"skip-auth-preflight" cfg:"skip_auth_preflight"`
-	ForceJSONErrors       bool     `flag:"force-json-errors" cfg:"force_json_errors"`
-	EncodeState           bool     `flag:"encode-state" cfg:"encode_state"`
-	AllowQuerySemicolons  bool     `flag:"allow-query-semicolons" cfg:"allow_query_semicolons"`
+	APIRoutes            []string `flag:"api-route" cfg:"api_routes"`
+	SkipAuthRegex        []string `flag:"skip-auth-regex" cfg:"skip_auth_regex"`
+	SkipAuthRoutes       []string `flag:"skip-auth-route" cfg:"skip_auth_routes"`
+	SkipJwtBearerTokens  bool     `flag:"skip-jwt-bearer-tokens" cfg:"skip_jwt_bearer_tokens"`
+	ExtraJwtIssuers      []string `flag:"extra-jwt-issuers" cfg:"extra_jwt_issuers"`
+	SkipProviderButton   bool     `flag:"skip-provider-button" cfg:"skip_provider_button"`
+	SkipAuthPreflight    bool     `flag:"skip-auth-preflight" cfg:"skip_auth_preflight"`
+	ForceJSONErrors      bool     `flag:"force-json-errors" cfg:"force_json_errors"`
+	EncodeState          bool     `flag:"encode-state" cfg:"encode_state"`
+	AllowQuerySemicolons bool     `flag:"allow-query-semicolons" cfg:"allow_query_semicolons"`
 
 	SignatureKey    string `flag:"signature-key" cfg:"signature_key"`
 	GCPHealthChecks bool   `flag:"gcp-healthchecks" cfg:"gcp_healthchecks"`
@@ -126,7 +125,6 @@ func NewFlagSet() *pflag.FlagSet {
 	flagSet.StringSlice("api-route", []string{}, "return HTTP 401 instead of redirecting to authentication server if token is not valid. Format: path_regex")
 	flagSet.Bool("skip-provider-button", false, "will skip sign-in-page to directly reach the next step: oauth/start")
 	flagSet.Bool("skip-auth-preflight", false, "will skip authentication for OPTIONS requests")
-	flagSet.Bool("ssl-insecure-skip-verify", false, "skip validation of certificates presented when using HTTPS providers")
 	flagSet.Bool("skip-jwt-bearer-tokens", false, "will skip requests that have verified JWT bearer tokens (default false)")
 	flagSet.Bool("force-json-errors", false, "will force JSON errors instead of HTTP error pages or redirects")
 	flagSet.Bool("encode-state", false, "will encode oauth state with base64")