+++ pkg/requests/builder.go
@@ -3,6 +3,7 @@ package requests
 import (
 	"context"
 	"fmt"
+	"golang.org/x/oauth2"
 	"io"
 	"net/http"
 )
@@ -16,6 +17,8 @@ type Builder interface {
 	WithBody(io.Reader) Builder
 	WithMethod(string) Builder
 	WithHeaders(http.Header) Builder
+	WithClientFromContext(context.Context) Builder
+	WithClient(*http.Client) Builder
 	SetHeader(key, value string) Builder
 	Do() Result
 }
@@ -27,11 +30,13 @@ type builder struct {
 	body     io.Reader
 	header   http.Header
 	result   *result
+	client   *http.Client
 }
 
 // New provides a new Builder for the given endpoint.
 func New(endpoint string) Builder {
 	return &builder{
+		client:   http.DefaultClient,
 		endpoint: endpoint,
 		method:   "GET",
 	}
@@ -59,6 +64,18 @@ func (r *builder) WithMethod(method string) Builder {
 // WithHeaders replaces the request header map with the given header map.
 func (r *builder) WithHeaders(header http.Header) Builder {
 	r.header = header
+	return r
+}
+
+func (r *builder) WithClientFromContext(ctx context.Context) Builder {
+	return r.WithContext(ctx).
+		WithClient(ctx.Value(oauth2.HTTPClient).(*http.Client))
+}
+
+func (r *builder) WithClient(client *http.Client) Builder {
+	if client != nil {
+		r.client = client
+	}
 	return r
 }
 
@@ -99,7 +116,7 @@ func (r *builder) do() Result {
 	}
 	req.Header = r.header
 
-	resp, err := http.DefaultClient.Do(req)
+	resp, err := r.client.Do(req)
 	if err != nil {
 		r.result = &result{err: fmt.Errorf("error performing request: %v", err)}
 		return r.result