+++ providers/gitlab.go
@@ -167,6 +167,7 @@ func (p *GitLabProvider) getUserinfo(ctx context.Context, s *sessions.SessionSta
 	var userinfo gitlabUserinfo
 	err := requests.New(userinfoURL.String()).
 		WithContext(ctx).
+		WithClient(p.Client).
 		SetHeader("Authorization", tokenTypeBearer+" "+s.AccessToken).
 		Do().
 		UnmarshalInto(&userinfo)
@@ -246,6 +247,7 @@ func (p *GitLabProvider) getProjectInfo(ctx context.Context, s *sessions.Session
 
 	err := requests.New(fmt.Sprintf("%s%s", endpointURL.String(), url.QueryEscape(project))).
 		WithContext(ctx).
+		WithClient(p.Client).
 		SetHeader("Authorization", tokenTypeBearer+" "+s.AccessToken).
 		Do().
 		UnmarshalInto(&projectInfo)