+++ providers/google.go
@@ -168,6 +168,7 @@ func (p *GoogleProvider) Redeem(ctx context.Context, redirectURL, code, codeVeri
 
 	err = requests.New(p.RedeemURL.String()).
 		WithContext(ctx).
+		WithClient(p.Client).
 		WithMethod("POST").
 		WithBody(bytes.NewBufferString(params.Encode())).
 		SetHeader("Content-Type", "application/x-www-form-urlencoded").
@@ -376,6 +377,7 @@ func (p *GoogleProvider) redeemRefreshToken(ctx context.Context, s *sessions.Ses
 
 	err = requests.New(p.RedeemURL.String()).
 		WithContext(ctx).
+		WithClient(p.Client).
 		WithMethod("POST").
 		WithBody(bytes.NewBufferString(params.Encode())).
 		SetHeader("Content-Type", "application/x-www-form-urlencoded").