+++ providers/internal_util.go
@@ -62,6 +62,7 @@ func validateToken(ctx context.Context, p Provider, accessToken string, header h
 
 	result := requests.New(endpoint).
 		WithContext(ctx).
+		WithClient(p.Data().Client).
 		WithHeaders(header).
 		Do()
 	if result.Error() != nil {