+++ providers/keycloak.go
@@ -75,6 +75,7 @@ func (p *KeycloakProvider) EnrichSession(ctx context.Context, s *sessions.Sessio
 
 	json, err := requests.New(profileURL).
 		WithContext(ctx).
+		WithClient(p.Client).
 		SetHeader("Authorization", tokenTypeBearer+" "+s.AccessToken).
 		Do().
 		UnmarshalSimpleJSON()