+++ providers/nextcloud.go
@@ -45,6 +45,7 @@ func (p *NextcloudProvider) EnrichSession(ctx context.Context, s *sessions.Sessi
 
 	json, err := requests.New(profileURL).
 		WithContext(ctx).
+		WithClient(p.Client).
 		SetHeader("Authorization", tokenTypeBearer+" "+s.AccessToken).
 		Do().
 		UnmarshalSimpleJSON()