+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Changes since v7.6.0
 
+- [#1838](https://github.com/oauth2-proxy/oauth2-proxy/pull/1838) Add configuration for provider client mTLS
+
 # V7.6.0
 
 ## Release Highlights